fixed some module/dependency mess by e.g: * removing SignalInformationPoint from ditto-interna-models-signal (as it pulled in several entity modules); moved funtionality of that interface to Command, Signal, etc. interfaces * moving HeaderTranslator from ditto-protocol to ditto-base-model * moving LikeHelper from ditto-rql-query to ditto-base-model * moved ConciergeForwarderActor and ShardRegions to ditto-edge-api * moved PolicyEnforcer from ditto-internal-utils-cache-loaders to ditto-policies-enforcement * created DittoSystemProperties inn ditto-base-model for commonly used system properties in order to avoid dependencies to models where not necessary * moved ThingCommandEnforcement, LiveSignalEnforcement, etc. from ditto-policies-enforcement to ditto-things-service where they should be located in the end
