+++ java/src/test/java/org/eclipse/ditto/client/live/TestConstants.java
@@ -20,6 +20,7 @@ import java.util.Collection;
 import org.eclipse.ditto.base.model.auth.AuthorizationContext;
 import org.eclipse.ditto.base.model.auth.AuthorizationSubject;
 import org.eclipse.ditto.base.model.auth.DittoAuthorizationContextType;
+import org.eclipse.ditto.base.model.common.DittoSystemProperties;
 import org.eclipse.ditto.base.model.headers.DittoHeaderDefinition;
 import org.eclipse.ditto.base.model.headers.DittoHeaders;
 import org.eclipse.ditto.json.JsonFactory;
@@ -38,7 +39,6 @@ import org.eclipse.ditto.things.model.ThingId;
 import org.eclipse.ditto.things.model.ThingLifecycle;
 import org.eclipse.ditto.things.model.ThingRevision;
 import org.eclipse.ditto.things.model.ThingsModelFactory;
-import org.eclipse.ditto.things.model.signals.commands.ThingCommandSizeValidator;
 import org.eclipse.ditto.things.model.signals.commands.exceptions.AttributeNotAccessibleException;
 import org.eclipse.ditto.things.model.signals.commands.exceptions.AttributeNotModifiableException;
 import org.eclipse.ditto.things.model.signals.commands.exceptions.AttributePointerInvalidException;
@@ -82,7 +82,7 @@ import org.eclipse.ditto.things.model.signals.commands.exceptions.ThingUnavailab
 public final class TestConstants {
 
     public static final long THING_SIZE_LIMIT_BYTES = Long.parseLong(
-            System.getProperty(ThingCommandSizeValidator.DITTO_LIMITS_THINGS_MAX_SIZE_BYTES, "-1"));
+            System.getProperty(DittoSystemProperties.DITTO_LIMITS_THINGS_MAX_SIZE_BYTES, "-1"));
 
     /**
      * A known correlation id for testing.
